+++ modules/tinymce/src/core/main/ts/delete/DeleteCommands.ts
@@ -5,7 +5,7 @@
  * For commercial licenses see https://www.tiny.cloud/
  */
 
-import { Cell } from '@ephox/katamari';
+import { Arr, Cell, Fun } from '@ephox/katamari';
 
 import Editor from '../api/Editor';
 import * as BlockBoundaryDelete from './BlockBoundaryDelete';
@@ -20,59 +20,39 @@ import * as MediaDelete from './MediaDelete';
 import * as Outdent from './Outdent';
 import * as TableDelete from './TableDelete';
 
-const nativeCommand = (editor: Editor, command: string): void => {
-  editor.getDoc().execCommand(command, false, null);
-};
+const findAction = (editor: Editor, caret: Cell<Text>, forward: boolean) =>
+  Arr.findMap([
+    Outdent.backspaceDelete,
+    CefDelete.backspaceDelete,
+    CaretBoundaryDelete.backspaceDelete,
+    (editor: Editor, forward: boolean) => InlineBoundaryDelete.backspaceDelete(editor, caret, forward),
+    BlockBoundaryDelete.backspaceDelete,
+    TableDelete.backspaceDelete,
+    ImageBlockDelete.backspaceDelete,
+    MediaDelete.backspaceDelete,
+    BlockRangeDelete.backspaceDelete,
+    InlineFormatDelete.backspaceDelete,
+  ], (item) => item(editor, forward));
 
 const deleteCommand = (editor: Editor, caret: Cell<Text>): void => {
-  if (Outdent.backspaceDelete(editor, false)) {
-    return;
-  } else if (CefDelete.backspaceDelete(editor, false)) {
-    return;
-  } else if (CaretBoundaryDelete.backspaceDelete(editor, false)) {
-    return;
-  } else if (InlineBoundaryDelete.backspaceDelete(editor, caret, false)) {
-    return;
-  } else if (BlockBoundaryDelete.backspaceDelete(editor, false)) {
-    return;
-  } else if (TableDelete.backspaceDelete(editor)) {
-    return;
-  } else if (ImageBlockDelete.backspaceDelete(editor, false)) {
-    return;
-  } else if (MediaDelete.backspaceDelete(editor, false)) {
-    return;
-  } else if (BlockRangeDelete.backspaceDelete(editor, false)) {
-    return;
-  } else if (InlineFormatDelete.backspaceDelete(editor, false)) {
-    return;
-  } else {
-    nativeCommand(editor, 'Delete');
-    DeleteUtils.paddEmptyBody(editor);
-  }
+  const result = findAction(editor, caret, false);
+
+  result.fold(
+    () => {
+      DeleteUtils.execDeleteCommand(editor);
+      DeleteUtils.paddEmptyBody(editor);
+    },
+    Fun.call
+  );
 };
 
 const forwardDeleteCommand = (editor: Editor, caret: Cell<Text>): void => {
-  if (CefDelete.backspaceDelete(editor, true)) {
-    return;
-  } else if (CaretBoundaryDelete.backspaceDelete(editor, true)) {
-    return;
-  } else if (InlineBoundaryDelete.backspaceDelete(editor, caret, true)) {
-    return;
-  } else if (BlockBoundaryDelete.backspaceDelete(editor, true)) {
-    return;
-  } else if (TableDelete.backspaceDelete(editor)) {
-    return;
-  } else if (ImageBlockDelete.backspaceDelete(editor, true)) {
-    return;
-  } else if (MediaDelete.backspaceDelete(editor, true)) {
-    return;
-  } else if (BlockRangeDelete.backspaceDelete(editor, true)) {
-    return;
-  } else if (InlineFormatDelete.backspaceDelete(editor, true)) {
-    return;
-  } else {
-    nativeCommand(editor, 'ForwardDelete');
-  }
+  const result = findAction(editor, caret, true);
+
+  result.fold(
+    () => DeleteUtils.execForwardDeleteCommand(editor),
+    Fun.call
+  );
 };
 
 const setup = (editor: Editor, caret: Cell<Text>): void => {