+++ modules/tinymce/src/core/main/ts/delete/InlineBoundaryDelete.ts
@@ -41,11 +41,10 @@ const hasOnlyTwoOrLessPositionsLeft = (elm: Node): boolean =>
       return CaretFinder.nextPosition(elm, normalizedFirstPos).forall((pos) => pos.isEqual(normalizedLastPos));
     }).getOr(true);
 
-const setCaretLocation = (editor: Editor, caret: Cell<Text>) => (location: BoundaryLocation.LocationAdt): boolean =>
-  BoundaryCaret.renderCaret(caret, location).exists((pos) => {
-    BoundarySelection.setCaretPosition(editor, pos);
-    return true;
-  });
+const setCaretLocation = (editor: Editor, caret: Cell<Text>) => (location: BoundaryLocation.LocationAdt): Optional<() => void> =>
+  BoundaryCaret.renderCaret(caret, location).map((pos) =>
+    () => BoundarySelection.setCaretPosition(editor, pos)
+  );
 
 const deleteFromTo = (editor: Editor, caret: Cell<Text>, from: CaretPosition, to: CaretPosition): void => {
   const rootNode = editor.getBody();
@@ -57,7 +56,8 @@ const deleteFromTo = (editor: Editor, caret: Cell<Text>, from: CaretPosition, to
 
     BoundaryLocation.readLocation(isInlineTarget, rootNode, CaretPosition.fromRangeStart(editor.selection.getRng()))
       .map(BoundaryLocation.inside)
-      .map(setCaretLocation(editor, caret));
+      .bind(setCaretLocation(editor, caret))
+      .each(Fun.call);
   });
 
   editor.nodeChanged();
@@ -68,12 +68,11 @@ const rescope = (rootNode: Node, node: Node): Node => {
   return parentBlock ? parentBlock : rootNode;
 };
 
-const backspaceDeleteCollapsed = (editor: Editor, caret: Cell<Text>, forward: boolean, from: CaretPosition): boolean => {
+const backspaceDeleteCollapsed = (editor: Editor, caret: Cell<Text>, forward: boolean, from: CaretPosition): Optional<() => void> => {
   const rootNode = rescope(editor.getBody(), from.container());
   const isInlineTarget = Fun.curry(InlineUtils.isInlineTarget, editor);
   const fromLocation = BoundaryLocation.readLocation(isInlineTarget, rootNode, from);
-
-  return fromLocation.bind((location) => {
+  const location = fromLocation.bind((location) => {
     if (forward) {
       return location.fold(
         Fun.constant(Optional.some(BoundaryLocation.inside(location))), // Before
@@ -89,42 +88,44 @@ const backspaceDeleteCollapsed = (editor: Editor, caret: Cell<Text>, forward: bo
         Fun.constant(Optional.some(BoundaryLocation.inside(location)))  // After
       );
     }
-  })
-    .map(setCaretLocation(editor, caret))
-    .getOrThunk(() => {
+  });
+
+  return location.map(setCaretLocation(editor, caret))
+    .getOrThunk((): Optional<() => void> => {
       const toPosition = CaretFinder.navigate(forward, rootNode, from);
       const toLocation = toPosition.bind((pos) => BoundaryLocation.readLocation(isInlineTarget, rootNode, pos));
 
       return Optionals.lift2(fromLocation, toLocation, () =>
-        InlineUtils.findRootInline(isInlineTarget, rootNode, from).exists((elm) => {
+        InlineUtils.findRootInline(isInlineTarget, rootNode, from).bind((elm) => {
           if (hasOnlyTwoOrLessPositionsLeft(elm)) {
-            DeleteElement.deleteElement(editor, forward, SugarElement.fromDom(elm));
-            return true;
+            return Optional.some(() => {
+              DeleteElement.deleteElement(editor, forward, SugarElement.fromDom(elm));
+            });
           } else {
-            return false;
+            return Optional.none();
           }
         })
-      ).orThunk(() => toLocation.bind((_) =>
+      ).getOrThunk(() => toLocation.bind(() =>
         toPosition.map((to) => {
-          if (forward) {
-            deleteFromTo(editor, caret, from, to);
-          } else {
-            deleteFromTo(editor, caret, to, from);
-          }
-
-          return true;
+          return () => {
+            if (forward) {
+              deleteFromTo(editor, caret, from, to);
+            } else {
+              deleteFromTo(editor, caret, to, from);
+            }
+          };
         })
-      )).getOr(false);
+      ));
     });
 };
 
-const backspaceDelete = (editor: Editor, caret: Cell<Text>, forward?: boolean): boolean => {
+const backspaceDelete = (editor: Editor, caret: Cell<Text>, forward?: boolean): Optional<() => void> => {
   if (editor.selection.isCollapsed() && Options.isInlineBoundariesEnabled(editor)) {
     const from = CaretPosition.fromRangeStart(editor.selection.getRng());
     return backspaceDeleteCollapsed(editor, caret, forward, from);
   }
 
-  return false;
+  return Optional.none();
 };
 
 export {