+++ modules/tinymce/src/core/main/ts/keyboard/FakeInputEvents.ts
@@ -5,6 +5,10 @@
  * For commercial licenses see https://www.tiny.cloud/
  */
 
+import { Fun } from '@ephox/katamari';
+
+import { EditorEvent } from 'tinymce/src/core/main/ts/api/PublicApi';
+
 import Editor from '../api/Editor';
 import { clone } from '../events/EventUtils';
 
@@ -12,31 +16,38 @@ interface SpecificsInput {
   data?: null | string;
 }
 
-const fireFakeInputEvent = (editor: Editor, inputType: string, specifics: SpecificsInput = {}) => {
-  const target = editor.getBody();
-  const overrides = {
-    bubbles: true,
-    composed: true,
-    data: null,
-    isComposing: false,
-    detail: 0,
-    view: null,
-    target,
-    currentTarget: target,
-    eventPhase: Event.AT_TARGET,
-    originalTarget: target,
-    explicitOriginalTarget: target,
-    isTrusted: false,
-    srcElement: target,
-    cancelable: false,
-    inputType
+const createAndFireInputEvent = (eventType: string) =>
+  (editor: Editor, inputType: string, specifics: SpecificsInput = {}): EditorEvent<InputEvent> => {
+    const target = editor.getBody();
+    const overrides = {
+      bubbles: true,
+      composed: true,
+      data: null,
+      isComposing: false,
+      detail: 0,
+      view: null,
+      target,
+      currentTarget: target,
+      eventPhase: Event.AT_TARGET,
+      originalTarget: target,
+      explicitOriginalTarget: target,
+      isTrusted: false,
+      srcElement: target,
+      cancelable: false,
+      preventDefault: Fun.noop,
+      inputType
+    };
+
+    const input = clone(new InputEvent(eventType));
+
+    return editor.fire(eventType, { ...input, ...overrides, ...specifics });
   };
 
-  const input = clone(new InputEvent('input'));
+const fireFakeInputEvent = createAndFireInputEvent('input');
 
-  editor.fire('input', { ...input, ...overrides, ...specifics });
-};
+const fireFakeBeforeInputEvent = createAndFireInputEvent('beforeinput');
 
 export {
-  fireFakeInputEvent
+  fireFakeInputEvent,
+  fireFakeBeforeInputEvent
 };